// client/modules/impower-game/data/enums/alignment.ts
export enum VerticalAlignment {
  Stretch = "Stretch",
  Middle = "Middle",
  Top = "Top",
  Bottom = "Bottom",
}

export enum HorizontalAlignment {
  Stretch = "Stretch",
  Center = "Center",
  Left = "Left",
  Right = "Right",
}

export enum TextAlignment {
  Center = "Center",
  Left = "Left",
  Right = "Right",
  Justify = "Justify",
}